requirements.txt
streamlit==1.29.0
pandas==2.1.3
openpyxl==3.1.2
numpy==1.26.2
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.0
bcrypt==4.1.1
python-jose==3.3.0
PyJWT==2.8.0
watchdog==3.0.0
python-multipart==0.0.6
typing-extensions==4.7.1
geopy==2.2.0
